Move ConstraintStoreLambdaVisitor in visitors folder
